legal-support/requirements.txt
Flask
pandas
openpyxl
google-cloud-storage
openai
config
